Don't check for mate in pos_check(), but check for ascension
The mate check was incomplete and is complicated to get right. Better
explicitly mark mated positions in the endgame database.

This commit also makes sure that the lions aren't in their opponents'
promotion zones.  Being there is not allowed as to simplify the endgame
database.

Incidentally this means that POS_GOTE is never returned, so remove it
from all files.