Remove promotion invariant from cohort map
It turns out that the invariant "If CHCK_S is promoted, then CHCK_G is
promoted, too" cannot be made to hold if we want CHCK_S and CHCK_G to
obey the ordering invariant, too.  So let's get rid of this invariant,
simplifying the map code in the process.