Add tabletest program
This program generates reduced tablebases containing only win/loss in 1
or undecided information. So far the results are inconclusive. Here is
the size of the resulting files (the ijk suffix indicates the loop
permutation, e.g. ijk indicates i being the outermost index, then j and
k being the innermost index) in kilobytes on disk (with zfs disk
compression enabled; the uncompressed files are all 291749376 bytes =
284911 kB in size).

52213 db.ikj
52193 db.kij
52017 db.ijk
51477 db.jik
47049 db.kji
46913 db.jki
 9225 db.kij.gz
 9225 db.ikj.gz
 8557 db.ijk.gz
 8133 db.jik.gz
 5925 db.kji.gz
 5857 db.jki.gz
 2281 db.ikj.zst
 2221 db.jki.zst
 2105 db.jik.zst
 1965 db.ijk.zst
 1801 db.jki.xz
 1757 db.ikj.xz
 1741 db.jik.xz
 1717 db.kij.zst
 1637 db.kji.zst
 1545 db.ijk.xz
 1457 db.kij.xz
 1349 db.kji.xz